+++ .eleventy.js
@@ -22,7 +22,7 @@ const path = require("node:path");
 const fs = require("node:fs");
 const yaml = require("js-yaml");
 const { DateTime } = require("luxon");
-const preWrapperPlugin = require("./src/_11ty/plugins/preWrapper.js");
+const preWrapperPlugin = require("./src/_11ty/plugins/pre-wrapper.js");
 
 //-----------------------------------------------------------------------------
 // Eleventy Config
